finbot/requirements.txt
slack-sdk==3.27.1
flask==3.0.0
requests==2.31.0
python-dotenv==1.0.0

